get sac training running

+++ lerobot/common/policies/sac/configuration_sac.py
@@ -51,7 +51,7 @@ class ActorNetworkConfig:
 @dataclass
 class PolicyConfig:
     use_tanh_squash: bool = True
-    log_std_min: int = -5
+    log_std_min: float = -5
     log_std_max: int = 2
     init_final: float = 0.05
